refactor: add dialog to enable sse
