+++ frontend/app/[team]/apps/[app]/syncing/page.tsx
@@ -1,33 +1,18 @@
 'use client'
 
 import GetAppSyncStatus from '@/graphql/queries/syncing/getAppSyncStatus.gql'
-import GetAppEnvironments from '@/graphql/queries/secrets/getAppEnvironments.gql'
-import InitAppSyncing from '@/graphql/mutations/syncing/initAppSync.gql'
-import GetEnvironmentKey from '@/graphql/queries/secrets/getEnvironmentKey.gql'
-import GetCfPages from '@/graphql/queries/syncing/cloudflare/getPages.gql'
-import { useLazyQuery, useMutation, useQuery } from '@apollo/client'
-import { unwrapEnvSecretsForUser, wrapEnvSecretsForServer } from '@/utils/environments'
-import { EnvironmentSyncType, EnvironmentType } from '@/apollo/graphql'
-import { Button } from '@/components/common/Button'
-import { OrganisationKeyring, cryptoUtils } from '@/utils/auth'
-import { Dialog, Transition } from '@headlessui/react'
-import clsx from 'clsx'
-import { useState, Fragment, useContext } from 'react'
-import { FaTimes, FaEyeSlash, FaEye, FaSync } from 'react-icons/fa'
-import { toast } from 'react-toastify'
-import { KeyringContext } from '@/contexts/keyringContext'
+import { useQuery } from '@apollo/client'
+import { EnvironmentSyncType } from '@/apollo/graphql'
+import { useContext } from 'react'
 import { organisationContext } from '@/contexts/organisationContext'
-import { useSession } from 'next-auth/react'
-import { Alert } from '@/components/common/Alert'
 import { SyncCard } from '@/components/syncing/SyncCard'
-import { ManageSyncDialog } from '@/components/syncing/ManageSyncDialog'
 import { SyncOptions } from '@/components/syncing/SyncOptions'
 import { useSearchParams } from 'next/navigation'
 import { userIsAdmin } from '@/utils/permissions'
+import { EnableSSEDialog } from '@/components/apps/EnableSSEDialog'
 
 export default function Syncing({ params }: { params: { team: string; app: string } }) {
   const { activeOrganisation: organisation } = useContext(organisationContext)
-  const { keyring } = useContext(KeyringContext)
 
   const searchParams = useSearchParams()
 
@@ -40,152 +25,6 @@ export default function Syncing({ params }: { params: { team: string; app: strin
 
   const activeUserIsAdmin = organisation ? userIsAdmin(organisation.role!) : false
 
-  const EnableSyncingDialog = () => {
-    const [enableSyncing, { loading }] = useMutation(InitAppSyncing)
-    const [getEnvKey] = useLazyQuery(GetEnvironmentKey)
-
-    const { data: appEnvsData } = useQuery(GetAppEnvironments, {
-      variables: {
-        appId: params.app,
-      },
-    })
-
-    const [isOpen, setIsOpen] = useState<boolean>(false)
-
-    const closeModal = () => {
-      setIsOpen(false)
-    }
-
-    const openModal = () => {
-      setIsOpen(true)
-    }
-
-    const handleEnableSyncing = async (e: { preventDefault: () => void }) => {
-      e.preventDefault()
-
-      const appEnvironments = appEnvsData.appEnvironments as EnvironmentType[]
-
-      const envKeyPromises = appEnvironments.map(async (env: EnvironmentType) => {
-        const { data: envKeyData } = await getEnvKey({
-          variables: {
-            envId: env.id,
-            appId: params.app,
-          },
-        })
-
-        const {
-          wrappedSeed: userWrappedSeed,
-          wrappedSalt: userWrappedSalt,
-          identityKey,
-        } = envKeyData.environmentKeys[0]
-
-        const { seed, salt } = await unwrapEnvSecretsForUser(
-          userWrappedSeed,
-          userWrappedSalt,
-          keyring!
-        )
-
-        const { wrappedSeed, wrappedSalt } = await wrapEnvSecretsForServer(
-          { seed, salt },
-          data.serverPublicKey
-        )
-
-        return {
-          envId: env.id,
-          identityKey,
-          wrappedSeed,
-          wrappedSalt,
-        }
-      })
-
-      const envKeyInputs = await Promise.all(envKeyPromises)
-
-      toast.promise(
-        enableSyncing({
-          variables: { appId: params.app, envKeys: envKeyInputs },
-          refetchQueries: [
-            {
-              query: GetAppSyncStatus,
-              variables: { appId: params.app },
-            },
-          ],
-        }),
-        {
-          pending: 'Enabling syncing...',
-          success: 'Syncing enabled for this App!',
-        }
-      )
-    }
-
-    return (
-      <>
-        <div className="flex items-center justify-center">
-          <Button variant="primary" onClick={openModal} title="Manage access">
-            <FaSync /> Enable Syncing
-          </Button>
-        </div>
-
-        <Transition appear show={isOpen} as={Fragment}>
-          <Dialog as="div" className="relative z-10" onClose={closeModal}>
-            <Transition.Child
-              as={Fragment}
-              enter="ease-out duration-300"
-              enterFrom="opacity-0"
-              enterTo="opacity-100"
-              leave="ease-in duration-200"
-              leaveFrom="opacity-100"
-              leaveTo="opacity-0"
-            >
-              <div className="fixed inset-0 bg-black/25 backdrop-blur-md" />
-            </Transition.Child>
-
-            <div className="fixed inset-0 overflow-y-auto">
-              <div className="flex min-h-full items-center justify-center p-4 text-center">
-                <Transition.Child
-                  as={Fragment}
-                  enter="ease-out duration-300"
-                  enterFrom="opacity-0 scale-95"
-                  enterTo="opacity-100 scale-100"
-                  leave="ease-in duration-200"
-                  leaveFrom="opacity-100 scale-100"
-                  leaveTo="opacity-0 scale-95"
-                >
-                  <Dialog.Panel className="w-full max-w-2xl transform overflow-hidden rounded-2xl bg-neutral-100 dark:bg-neutral-900 p-6 text-left align-middle shadow-xl transition-all">
-                    <Dialog.Title as="div" className="flex w-full justify-between">
-                      <h3 className="text-lg font-medium leading-6 text-black dark:text-white ">
-                        Enable syncing for this app
-                      </h3>
-
-                      <Button variant="text" onClick={closeModal}>
-                        <FaTimes className="text-zinc-900 dark:text-zinc-400 hover:text-zinc-700 dark:hover:text-zinc-300" />
-                      </Button>
-                    </Dialog.Title>
-
-                    <form className={clsx('space-y-6 p-4')} onSubmit={handleEnableSyncing}>
-                      <Alert variant="warning" icon={true}>
-                        Enabling syncing for this App will allow the server to access secrets in all
-                        environments.
-                      </Alert>
-
-                      <div className="flex items-center gap-4">
-                        <Button variant="secondary" type="button" onClick={closeModal}>
-                          Cancel
-                        </Button>
-                        <Button variant="primary" type="submit" isLoading={loading}>
-                          Enable
-                        </Button>
-                      </div>
-                    </form>
-                  </Dialog.Panel>
-                </Transition.Child>
-              </div>
-            </div>
-          </Dialog>
-        </Transition>
-      </>
-    )
-  }
-
   return (
     <div className="w-full space-y-8 pt-8 text-black dark:text-white">
       {data?.syncEnabled === false && (
@@ -196,7 +35,7 @@ export default function Syncing({ params }: { params: { team: string; app: strin
               Syncing is not yet enabled for this app. Click the button below to enable syncing.
             </div>
           </div>
-          <EnableSyncingDialog />
+          <EnableSSEDialog appId={params.app} />
         </div>
       )}
       {data?.syncEnabled === true && (
